Assignment_2/NLP_Assignment_2-Emotion_Classification/requirements.txt
pandas==2.2.1
PyYAML==6.0.1
